requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
PyMySQL==1.1.0
openai==1.3.5
python-dotenv==1.0.0
django-cors-headers==4.3.1
cryptography==41.0.7
sqlparse==0.4.4
boto3==1.29.7